+++ demos/database.php
@@ -39,7 +39,8 @@ class Stat extends \atk4\data\Model
         parent::init();
 
         $this->addFields(['project_name', 'project_code'], ['type'=>'string']);
-        $this->addField('description', ['ui'=>['form'=>['FormField/TextArea', 'rows'=>5]]]);
+        //$this->addField('description', ['ui'=>['form'=>['FormField/TextArea', 'rows'=>5]]]);
+        $this->addField('description', ['type'=>'text']);
         $this->addField('client_name', ['type'=>'string']);
         $this->addField('client_address', ['type'=>'string', 'ui'=>['form'=>[new \atk4\ui\FormField\Textarea(), 'rows'=>4]]]);
 
@@ -79,7 +80,7 @@ class Stat extends \atk4\data\Model
         $this->addFields(['start_date', 'finish_date'], ['type'=>'date']);
         $this->addField('finish_time', ['type'=>'time']);
 
-        $this->addFields(['created', 'updated'], ['type'=>'datetime']);
+        $this->addFields(['created', 'updated'], ['type'=>'datetime', 'ui'=>['form'=>['FormField/Line', 'disabled'=>true]]]);
     }
 }