+++ src/FormField/Dropdown.php
@@ -13,6 +13,8 @@ class Dropdown extends Input
 
     public $values = [];
 
+    public $empty = '...';
+
     public function init()
     {
         parent::init();
@@ -27,21 +29,39 @@ class Dropdown extends Input
         $value = isset($this->field) ? $this->app->ui_persistence->typecastSaveField($this->field, $this->field->get()) : $this->content ?: '';
 
         $options = [];
-        foreach ($this->values as $key=>$val) {
-            $item = ['option', 'value'=>(string) $key, $val];
-            if ($value == $val) {
-                $item['selected'] = true;
-            }
+        if ($this->empty) {
+            $item = ['option', 'value'=>'', $this->empty];
             $options[] = $item;
         }
 
+        if (isset($this->model)) {
+            foreach ($this->model as $key=>$row) {
+                $title = $row[$row->title_field];
+
+                $item = ['option', 'value'=>(string) $key, $title];
+                if ($value == $key) {
+                    $item['selected'] = true;
+                }
+                $options[] = $item;
+            }
+        } else {
+            foreach ($this->values as $key=>$val) {
+                $item = ['option', 'value'=>(string) $key, $val];
+                if ($value == $key) {
+                    $item['selected'] = true;
+                }
+                $options[] = $item;
+            }
+        }
+
         return $this->app->getTag('select', [
+            'class'      => 'fluid search selection',
             'name'       => $this->short_name,
             'type'       => $this->inputType,
             'rows'       => $this->rows,
             'placeholder'=> $this->placeholder,
             'id'         => $this->id.'_input',
-        ], [$options]
+        ], [[$options]]
        //
     );
         //return '<input name="'.$this->short_name.'" type="'.$this->inputType.'" placeholder="'.$this->placeholder.'" id="'.$this->id.'_input"/>';