+++ src/TableColumn/Generic.php
@@ -73,20 +73,8 @@ class Generic
         return $this;
     }
 
-    /**
-     * Returns a suitable cell tag with the supplied value. Applies modifiers
-     * added through addClass and setAttr.
-     *
-     * @param string $tag
-     * @param string $position
-     * @param string $value
-     *
-     * @return string
-     */
-    public function getTag($tag, $position, $value)
+    public function getTagAttributes($position, $attr = [])
     {
-        $attr = [];
-
         // "all" applies on all positions
         if (isset($this->attr['all'])) {
             $attr = array_merge_recursive($attr, $this->attr['all']);
@@ -96,6 +84,23 @@ class Generic
         if (isset($this->attr[$position])) {
             $attr = array_merge_recursive($attr, $this->attr[$position]);
         }
+
+        return $attr;
+    }
+
+    /**
+     * Returns a suitable cell tag with the supplied value. Applies modifiers
+     * added through addClass and setAttr.
+     *
+     * @param string $position - 'head', 'body' or 'tail'
+     * @param string $value    - what is inside? either html or array defining HTML structure, see App::getTag help
+     * @param array  $attr     - extra attributes to apply on the tag
+     *
+     * @return string
+     */
+    public function getTag($position, $value, $attr = [])
+    {
+        $attr = $this->getTagAttributes($position, $attr);
 
         if (isset($attr['class'])) {
             $attr['class'] = implode(' ', $attr['class']);
@@ -108,19 +113,37 @@ class Generic
      * Provided with a field definition (from a model) will return a header
      * cell, fully formatted to be included in a Table. (<th>).
      *
-     * Potentially may include elements for sorting.
-     *
      * @param \atk4\data\Field $f
      *
      * @return string
      */
-    public function getHeaderCellHTML(\atk4\data\Field $f = null)
+    public function getHeaderCellHTML(\atk4\data\Field $f = null, $value = null)
     {
         if ($f === null) {
-            return $this->getTag('th', 'head', '');
+            return $this->getTag('head', '', $this->table->sortable ? ['class'=>['disabled']] : []);
         }
 
-        return $this->getTag('th', 'head', $f->getCaption());
+        // If table is being sorted by THIS column, set the proper class
+        $attr = [];
+        if ($this->table->sortable) {
+            $attr['data-column'] = $f->short_name;
+
+            if ($this->table->sort_by === $f->short_name) {
+                $attr['class'][] = 'sorted '.$this->table->sort_order;
+
+                if ($this->table->sort_order === 'ascending') {
+                    $attr['data-column'] = '-'.$f->short_name;
+                } elseif ($this->table->sort_order === 'descending') {
+                    $attr['data-column'] = '';
+                }
+            }
+        }
+
+        return $this->getTag(
+            'head',
+            $f->getCaption(),
+            $attr
+        );
     }
 
     /**
@@ -133,7 +156,7 @@ class Generic
      */
     public function getTotalsCellHTML(\atk4\data\Field $f, $value)
     {
-        return $this->getTag('th', 'foot', $this->app->ui_persistence->typecastSaveField($f, $value));
+        return $this->getTag('foot', $this->app->ui_persistence->typecastSaveField($f, $value));
     }
 
     /**
@@ -152,13 +175,33 @@ class Generic
      *
      * @return string
      */
-    public function getDataCellHTML(\atk4\data\Field $f = null)
+    public function getDataCellHTML(\atk4\data\Field $f = null, $extra_tags = [])
     {
-        if ($f === null) {
-            return $this->getTag('td', 'body', '{$c_'.$this->short_name.'}');
-        }
+        return $this->getTag('body', [$this->getDataCellTemplate($f)], $extra_tags);
+    }
 
-        return $this->getTag('td', 'body', '{$'.$f->short_name.'}');
+    /**
+     * Provided with a field definition will return a string containing a "Template"
+     * that would produce CONTENS OF <td> cell when rendered. Example output:.
+     *
+     *   <b>{$name}</b>
+     *
+     * The tag that corresponds to the name of the field (e.g. {$name}) may be substituted
+     * by another template returned by getDataCellTemplate when multiple formatters are
+     * applied to the same column. The first one to be applied is executed first, then
+     * a subsequent ones are executed.
+     *
+     * @param \atk4\data\Field $f
+     *
+     * @return string
+     */
+    public function getDataCellTemplate(\atk4\data\Field $f = null)
+    {
+        if ($f) {
+            return '{$'.$f->short_name.'}';
+        } else {
+            return '{_$'.$this->short_name.'}';
+        }
     }
 
     /**