+++ src/TableColumn/Money.php
@@ -13,27 +13,34 @@ class Money extends Generic
     // overrides
     public $attr = ['all'=>['class'=>['right aligned single line']]];
 
-    public function getDataCellHTML(\atk4\data\Field $f = null)
+    public function getTagAttributes($position, $attr = [])
+    {
+        $attr = array_merge_recursive($attr, ['class'=>['{$_'.$this->short_name.'_class}']]);
+
+        return parent::getTagAttributes($position, $attr);
+    }
+
+    public function getDataCellHTML(\atk4\data\Field $f = null, $extra_tags = [])
     {
         if (!isset($f)) {
             throw new Exception(['Money column requires a field']);
         }
 
-        return $this->app->getTag(
-            'td',
-            ['class'=> '{$_'.$f->short_name.'_money}'],
-            '{$'.$f->short_name.'}'
+        return $this->getTag(
+            'body',
+            '{$'.$f->short_name.'}',
+            $extra_tags
         );
     }
 
     public function getHTMLTags($row, $field)
     {
         if ($field->get() < 0) {
-            return ['_'.$field->short_name.'_money'=>'right aligned single line negative'];
+            return ['_'.$this->short_name.'_class'=>'negative'];
         } elseif (!$this->show_zero_values && $field->get() == 0) {
-            return ['_'.$field->short_name.'_money'=>'right aligned single line', $field->short_name=>'-'];
+            return ['_'.$this->short_name.'_class'=>'', $field->short_name=>'-'];
         }
 
-        return ['_'.$field->short_name.'_money'=>'right aligned single line'];
+        return ['_'.$this->short_name.'_class'=>''];
     }
 }