+++ tests/jsIntegrationTest.php
@@ -84,9 +84,7 @@ $(function() {
 
         $this->assertEquals('<script>
 $(function() {
-  $("#b1").on("click",function(event) {
-    event.preventDefault();
-    event.stopPropagation();
+  $("#b1").on("click",function() {
     $("#b2").hide();
   });
 })</script>', $bb->getJS());